+++ packages/javascript/lib/node/os/getArch.ts
@@ -1,0 +1,5 @@
+import { arch } from 'node:os';
+
+export function getArch(): string {
+	return arch();
+}