+++ packages/javascript/lib/node/os/getEndianness.ts
@@ -1,0 +1,5 @@
+import { endianness } from 'node:os';
+
+export function getEndianness(): 'BE' | 'LE' {
+	return endianness();
+}