+++ packages/javascript/lib/node/os/getKernelVersion.ts
@@ -1,0 +1,5 @@
+import { release } from 'node:os';
+
+export function getKernelVersion(): string {
+	return release() || 'Unknown';
+}